requirements.txt
pdfplumber
python-docx
Pillow
fastapi
uvicorn
python-multipart
pytest
httpx
# pytesseract  # Optional dependency
